change: load anonymization functions from classpath

+++ outis-core/src/main/scala/io/saagie/outis/core/job/AnonymizationJob.scala
@@ -1,17 +1,23 @@
-package io.saagie.job
+package io.saagie.outis.core.job
+
+import java.lang.reflect.Method
 
 import com.databricks.spark.avro._
 import io.saagie.model._
-import io.saagie.outis.core.anonymize.AnonymizeString
+import io.saagie.outis.core.anonymize.AnonymizationException
 import io.saagie.util.HdfsUtils
 import org.apache.hadoop.fs.Path
+import org.apache.spark.sql.types.StringType
 import org.apache.spark.sql.{DataFrame, SparkSession}
 
-case class Anonymize(dataset: DataSet)(implicit spark: SparkSession) {
+import scala.reflect.internal.util.ScalaClassLoader
+import scala.util.{Failure, Success, Try}
+
+case class AnonymizationJob(dataset: DataSet, outisConf: OutisConf = OutisConf())(implicit spark: SparkSession) {
 
   import org.apache.spark.sql.functions.col
 
-  def anonymise(): Unit = {
+  def anonymize(): Unit = {
 
     dataset match {
       case d: HiveDataSet => anonymiseFromHive(d)
@@ -44,6 +50,17 @@ case class Anonymize(dataset: DataSet)(implicit spark: SparkSession) {
     HdfsUtils(dataset.hdfsUrl).deleteFiles(List(new Path(dataset.hdfsUrl)))
   }
 
+  private def getStringAnonymization: Either[AnonymizationException, Method] = {
+    ScalaClassLoader(getClass.getClassLoader).tryToLoadClass(outisConf.getClassFor(OutisConf.ANONYMIZER_STRING)) match {
+      case Some(x: Class[_]) =>
+        Try(x.getDeclaredMethod(outisConf.getMethodFor(OutisConf.ANONYMIZER_STRING), outisConf.getParameterClassesFor(OutisConf.ANONYMIZER_STRING): _*)) match {
+          case Success(m) => Right(m)
+          case Failure(e) => Left(AnonymizationException(s"Impossible to load method: ${outisConf.getMethodFor(OutisConf.ANONYMIZER_STRING)}", e))
+        }
+      case None => Left(AnonymizationException(s"Anonymization class not found: ${outisConf.getClassFor(OutisConf.ANONYMIZER_STRING)}"))
+    }
+  }
+
   private def anonymiseFromHive[T <: HiveDataSet](dataset: T) = {
 
     val table = dataset.table
@@ -52,22 +69,40 @@ case class Anonymize(dataset: DataSet)(implicit spark: SparkSession) {
 
     val df: DataFrame = spark.sql(s"SELECT * FROM $table")
 
-    val anonymize = spark.sqlContext.udf.register("anonymize", s => AnonymizeString.setToX(s))
+    val anonymizeString = getStringAnonymization.right.map(method => spark.sqlContext.udf.register("anonymizeString",
+      s =>
+        method.invoke(null, outisConf.getParameters(OutisConf.ANONYMIZER_STRING).map {
+          case ColumnValue => s
+          case x => x
+        }: _*)
+    ))
 
-    val columnsNonAnonymized = df.columns.filter(c => !(dataset.columnsToAnonymise contains c))
-    df.select(columnsNonAnonymized.map(c => col(c).alias(c)).union(dataset.columnsToAnonymise.map(c => anonymize(col(c).alias(c)))): _*)
+    if (anonymizeString.isRight) {
+      val stringAnonimyzer = anonymizeString.right.get
+      val columnsNonAnonymized = df.columns.filter(c => !(dataset.columnsToAnonymise contains c))
+      df.select(
+        columnsNonAnonymized.map(c => col(c).alias(c))
+          .union(dataset.columnsToAnonymise
+            .map(c => {
+              df.schema(c).dataType match {
+                case StringType => stringAnonimyzer(col(c)).alias(c)
+                case _ => col(c).alias(c)
+              }
+            })
+          ): _*)
 
-    df.createOrReplaceTempView(sparkTmpTable)
+      df.createOrReplaceTempView(sparkTmpTable)
 
-    val options: String = dataset match {
-      case d: TextFileHiveDataset => s"OPTIONS(fileFormat '${dataset.storageFormat.toString}', fieldDelim '${d.fieldDelimiter}', escapeDelim '${d.escapeDelimiter}', collectionDelim '${d.collectionDelimiter}', mapkeyDelim '${d.mapKeyDelimiter}', lineDelim '${d.lineDelimiter}')"
-      case _ => s"OPTIONS(fileFormat '${dataset.storageFormat.toString}')"
+      val options: String = dataset match {
+        case d: TextFileHiveDataset => s"OPTIONS(fileFormat '${dataset.storageFormat.toString}', fieldDelim '${d.fieldDelimiter}', escapeDelim '${d.escapeDelimiter}', collectionDelim '${d.collectionDelimiter}', mapkeyDelim '${d.mapKeyDelimiter}', lineDelim '${d.lineDelimiter}')"
+        case _ => s"OPTIONS(fileFormat '${dataset.storageFormat.toString}')"
+      }
+
+      val createTmpTable = s"CREATE TABLE $tmpTable AS SELECT * FROM  $sparkTmpTable USING hive $options"
+      spark.sql(createTmpTable)
+
+      val dropTable = s"DROP TABLE $table"
+      spark.sql(dropTable)
     }
-
-    val createTmpTable = s"CREATE TABLE $tmpTable AS SELECT * FROM  $sparkTmpTable USING hive $options"
-    spark.sql(createTmpTable)
-
-    val dropTable = s"DROP TABLE $table"
-    spark.sql(dropTable)
   }
 }
